+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Each release is linked to a git commit.
 - Constellation and LEO builder
 - Local Time at Ascending Node computation module
 - Real adaptive step size for KeplerNum propagator
+- Runge-Kutta-Fehlberg method for KeplerNum
 
 ### Modified